@@ -7,20 +7,26 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@babel/core": "^7.3.4",
-    "axios": "^0.18.0",
+    "@babel/core": "^7.7.7",
+    "axios": "^0.18.1",
     "babel": "^6.23.0",
-    "expo": "^32.0.6",
-    "expo-cli": "^2.11.2",
-    "expo-core": "^2.0.1",
-    "native-base": "^2.8.2",
-    "react": "^16.5.0",
-    "react-native": "^0.57.1",
-    "react-native-vector-icons": "^6.3.0",
-    "react-navigation": "^3.3.0"
+    "dev-tools": "^0.1.6",
+    "expo": "^36.0.2",
+    "expo-av": "~8.0.0",
+    "expo-cli": "^2.21.2",
+    "expo-core": "^2.0.2",
+    "expo-font": "~8.0.0",
+    "expo-secure-store": "^8.0.0",
+    "native-base": "^2.13.8",
+    "react": "^16.12.0",
+    "react-native": "^0.61.5",
+    "react-native-gesture-handler": "^1.5.3",
+    "react-native-vector-icons": "^6.6.0",
+    "react-native-webview": "7.4.3",
+    "react-navigation": "^3.11.0"
   },
   "devDependencies": {
-    "babel-preset-expo": "^5.0.0",
+    "babel-preset-expo": "^5.2.0",
     "schedule": "^0.4.0"
   },
   "private": true
