@@ -17,7 +17,7 @@ export default class Playlist extends React.Component {
     }
   }
 
-  componentWillMount = () => {
+  UNSAFE_componentWillMount = () => {
     this.setState({userId: this.props.navigation.getParam('userId')})
   }
   componentDidMount = () => {
